+++ code/bash/mothurPrecluster.sh
@@ -32,7 +32,9 @@ echo PROGRESS: Creating contigs for all the samples
 mkdir -p "${TMP}"/
 
 # Removing old files if they exist
-rm "${OUTDIR}"/*
+if [ -n "$(ls -A "${OUTDIR}")" ]; then
+	rm "${OUTDIR}"/*
+fi
 
 mothur "#make.contigs(file="${FILESFILE}", outputdir="${TMP}"/, processors="${NPROC}");
 	screen.seqs(fasta=current, group=current, maxambig=0, maxlength=275, maxhomop=8);
